Add support to use Self-Signed CA certificate, Fix syntax, Update Vagrant file to be more dynamic

+++ Playground/Helm/Cert-Manager/cert-manager.sh
@@ -1,5 +1,4 @@
 #! /bin/bash
-set -e
 
 # Dependecy checks
 if [[ ! $1 ]]; then
@@ -10,16 +9,32 @@ elif [[ $1 != "install" ]] && [[ $1 != "uninstall" ]]; then
   exit 1
 fi
 
-if ! which base64 &> /dev/null; then echo "Helm does not exists, please install it and re-run $0"; exit 1; fi
-if ! which helm &> /dev/null; then echo "Helm does not exists, please install it and re-run $0"; exit 1; fi
-if ! which kubectl &> /dev/null; then echo "Kubectl does not exists, please install it and re-run $0"; exit 1; fi
-if ! which openssl &> /dev/null; then echo "OpenSSL does not exists, please install it and re-run $0"; exit 1; fi
+if ! which base64 &> /dev/null; then echo -e "`date +"%d-%m-%y %H:%M:%S"`\tBase64 is missing!" | tee -a /var/log/cert-manager.log; exit 2; fi
+if ! which helm &> /dev/null; then echo -e "`date +"%d-%m-%y %H:%M:%S"`\tHelm is missing!" | tee -a /var/log/cert-manager.log; exit 2; fi
+if ! which kubectl &> /dev/null; then echo -e "`date +"%d-%m-%y %H:%M:%S"`\tKubectl is missing!" | tee -a /var/log/cert-manager.log; exit 2; fi
+if ! which openssl &> /dev/null; then echo -e "`date +"%d-%m-%y %H:%M:%S"`\tOpenSSL is missing!" | tee -a /var/log/cert-manager.log; exit 2; fi
+
+NODES_COUNT=$(kubectl get nodes -o json | jq -r '.items[] | select(.spec.taints|not) | select(.status.conditions[].reason=="KubeletReady" and .status.conditions[].status=="True") | .metadata.name' | wc -l)
+if [ ! $NODES_COUNT -gt 0 ]; then
+  echo -e "`date +"%d-%m-%y %H:%M:%S"`\tCert-Manager setup requires at least 1 node, currently there are ${NODES_COUNT} registered, please enlarge your K8s cluster" | tee -a /var/log/cert-manager.log
+  exit 3
+fi
 
 # Variables
 CA_NAME='k8s-playground-ca'
 NAMESPACE='cert-manager'
 
 install() {
+  # Protect the script from re-run
+  if [ ! -f /var/run/cert-manager.pid ]; then
+    echo $$ > /var/run/cert-manager.pid
+  else
+    echo -e "`date +"%d-%m-%y %H:%M:%S"`\tCert-Manager setup is already running in the background..." | tee -a /var/log/cert-manager.log
+    echo -e "`date +"%d-%m-%y %H:%M:%S"`\tUse \`kubectl get events -n cert-manager -w\` to check the setup progress..." | tee -a /var/log/cert-manager.log
+    exit 4
+  fi
+
+  echo -e "`date +"%d-%m-%y %H:%M:%S"`\tInstalling Cert-Manager from Helm..." | tee -a /var/log/cert-manager.log
   helm upgrade --install cert-manager cert-manager \
     --repo https://charts.jetstack.io \
     --create-namespace \
@@ -27,7 +42,8 @@ install() {
     --set installCRDs=true
   kubectl wait -n ${NAMESPACE} pod -l app.kubernetes.io/name=cert-manager --for condition=Ready --timeout=60s
 
-  if ! kubectl -n ssl-ready get secret/${CA_NAME}-secret > /dev/null 2>&1 && ! kubectl -n ${NAMESPACE} get secret/${CA_NAME}-secret > /dev/null 2>&1; then
+  if ! kubectl -n ssl-ready get secret/${CA_NAME}-secret &> /dev/null && ! kubectl -n ${NAMESPACE} get secret/${CA_NAME}-secret &> /dev/null; then
+    echo -e "`date +"%d-%m-%y %H:%M:%S"`\tGenerating self-Signed CA autority certificate..." | tee -a /var/log/cert-manager.log
     CA_LIFE_TIME='3650'
     ROOT_CA_PATH='root-ca'
     
@@ -51,14 +67,16 @@ data:
   tls.crt: $(cat ${ROOT_CA_PATH}/root-ca.crt | base64 -w 0)
   tls.key: $(cat ${ROOT_CA_PATH}/root-ca.key | base64 -w 0)
 EOF
-  elif kubectl -n ssl-ready get secret/${CA_NAME}-secret > /dev/null 2>&1 && ! kubectl -n ${NAMESPACE} get secret/${CA_NAME}-secret > /dev/null 2>&1; then
-  CRT_DATA=`kubectl -n ssl-ready get secret/${CA_NAME} -o jsonpath='{.data.tls\.crt}'`
-  KEY_DATA=`kubectl -n ssl-ready get secret/${CA_NAME} -o jsonpath='{.data.tls\.key}'`
+
+  elif kubectl -n ssl-ready get secret/${CA_NAME}-secret &> /dev/null && ! kubectl -n ${NAMESPACE} get secret/${CA_NAME}-secret &> /dev/null; then
+    echo -e "`date +"%d-%m-%y %H:%M:%S"`\tImporting an already existing CA autority certificate..." | tee -a /var/log/cert-manager.log
+  CRT_DATA=`kubectl -n ssl-ready get secret/${CA_NAME}-secret -o jsonpath='{.data.tls\.crt}'`
+  KEY_DATA=`kubectl -n ssl-ready get secret/${CA_NAME}-secret -o jsonpath='{.data.tls\.key}'`
 cat <<EOF | kubectl apply -f -
 apiVersion: v1
 kind: Secret
 metadata:
-  name: ${CA_NAME}
+  name: ${CA_NAME}-secret
   namespace: ${NAMESPACE}
 type: Opaque
 data:
@@ -76,16 +94,18 @@ spec:
   ca:
     secretName: ${CA_NAME}-secret
 EOF
-  kubectl wait ClusterIssuer ${CA_NAME}-issuer --for condition=Ready --timeout=60s
 
   echo "1st note: Make sure you distribute the new CA autorithy certificate to all cluster nodes"
   echo "2nd note: Make sure you distribute the new CA autorithy certificate to all systems\applications that may require access to services signed by this ROOT CA"
+
+  if [ -f /etc/cron.d/cert-manager-setup ]; then rm -f /etc/cron.d/cert-manager-setup; fi
+  if [ -f /var/run/cert-manager.pid ]; then rm -f /var/run/cert-manager.pid; fi
 }
 
 uninstall() {
-  kubectl delete ClusterIssuer/${CA_NAME}-issuer
-  kubectl delete -n ${NAMESPACE} Secret/${CA_NAME}-secret
-  helm uninstall cert-manager cert-manager --namespace ${NAMESPACE}
+  if kubectl get ClusterIssuer/${CA_NAME}-issuer &> /dev/null; then kubectl delete ClusterIssuer/${CA_NAME}-issuer; fi
+  if kubectl get -n ${NAMESPACE} Secret/${CA_NAME}-secret &> /dev/null; then kubectl delete -n ${NAMESPACE} Secret/${CA_NAME}-secret; fi
+  if kubectl get ns ${NAMESPACE} &> /dev/null; then helm uninstall cert-manager cert-manager --namespace ${NAMESPACE}; fi
   kubectl delete ns ${NAMESPACE}
 
   echo "1st note: Please remove this CA autorithy certificate from all cluster nodes"
